+++ example/spa_with_states/web/src/const/Action.dart
@@ -1,0 +1,13 @@
+class Action {
+  static final String INITIALIZE = "action_initialize";
+
+  static final String LOGIN_PAGE_BUTTON_INDEX_CLICKED = "action_login_page_button_index_clicked";
+  static final String LOGIN_PAGE_BUTTON_GALLERY_CLICKED = "action_login_page_button_gallery_clicked";
+
+  static final String GALLERY_PAGE_BUTTON_INDEX_CLICKED = "action_gallery_page_button_index_clicked";
+  static final String GALLERY_PAGE_BUTTON_EXIT_CLICKED = "action_gallery_page_button_exit_clicked";
+
+  static final String INDEX_PAGE_BUTTON_LOGIN_CLICKED = "action_index_page_button_login_clicked";
+
+  static final String SIGNOUT_PAGE_TIMER_EXPIRED = "action_signout_page_timer_expired";
+}