+++ CHANGELOG.md
@@ -1,5 +1,9 @@
+## 0.9.1:
+- SPA with states example
+- StateAction and StateMeta
+
 ## 0.9.0:
-- API change: DartMachine -> States
+- API change: States -> States
 
 ## 0.1.9
 - Internal files renamed no effect on functionality